Move util.ts to utils
