+++ frontend/src/features/data_entry_choice/components/PollingStationSelector.tsx
@@ -5,8 +5,8 @@ import { t, tx } from "@/lib/i18n";
 import { IconError, IconWarning } from "@/lib/icon";
 import { cn, removeLeadingZeros } from "@/lib/util";
 
+import { PollingStationUserStatus, PollingStationWithStatus } from "../utils/util";
 import cls from "./DataEntryChoice.module.css";
-import { PollingStationUserStatus, PollingStationWithStatus } from "./util";
 
 export interface PollingStationSelectorProps {
   pollingStationNumber: string;
