+++ frontend/src/features/data_entry_choice/utils/util.ts
